refactor : 패키지 의미 명확하게 이름 변경
