--- src/main/java/racingStep3/RacingApplication.java
@@ -1,13 +1,0 @@
-package racingStep3;
-
-import racingStep3.controller.RacingCarController;
-import racingStep3.view.InputView;
-
-public class RacingApplication {
-    public static void main(String[] args) {
-
-        RacingCarController racingCarController = new RacingCarController(new InputView());
-        racingCarController.execute();
-
-    }
-}
