--- src/main/java/racingStep3/service/strategy/MoveRule.java
@@ -1,6 +1,0 @@
-package racingStep3.service.strategy;
-
-@FunctionalInterface
-public interface MoveRule {
-    boolean moveOrNot();
-}
