--- src/main/java/racingStep3/service/util/MoveRuleUtil.java
@@ -1,14 +1,0 @@
-package racingStep3.service.util;
-
-import racingStep3.service.strategy.MoveRule;
-
-public class MoveRuleUtil implements MoveRule {
-
-    private static final int MIN_RANDOM = 4;
-    private static final int MAX_RANDOM = 10;
-
-    @Override
-    public boolean moveOrNot() {
-        return MIN_RANDOM <= RacingUtil.getRandomValue(MAX_RANDOM);
-    }
-}
