+++ src/main/java/racingcar/service/strategy/MoveRule.java
@@ -1,0 +1,6 @@
+package racingcar.service.strategy;
+
+@FunctionalInterface
+public interface MoveRule {
+    boolean moveOrNot();
+}
