--- src/test/java/racingStep1/SetTest.java
@@ -1,43 +1,0 @@
-package racingStep1;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
-
-import java.util.HashSet;
-import java.util.Set;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class SetTest {
-    private Set<Integer> numbers;
-
-    @BeforeEach
-    void setUp() {
-        numbers = new HashSet<>();
-        numbers.add(1);
-        numbers.add(1);
-        numbers.add(2);
-        numbers.add(3);
-    }
-
-    // Test Case 구현
-    @Test
-    void sizeCheck() {
-        assertThat(numbers.size()).isEqualTo(3);
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = {1, 2, 3})
-    void containCheck(int input) {
-        assertThat(numbers.contains(input)).isTrue();
-    }
-
-    @ParameterizedTest
-    @CsvSource(value = {"1:true", "2:true", "3:true", "4:false", "5:false"}, delimiter = ':')
-    void extendCheck(int input, boolean expect) {
-        assertThat(numbers.contains(input)).isEqualTo(expect);
-    }
-}
